+++ tests/test_app/TestApp/Application.php
@@ -18,6 +18,8 @@ class Application extends BaseApplication
      */
     public function bootstrap(): void
     {
+        parent::bootstrap();
+
         $this->addPlugin(MeTools::class);
     }